transaction in pwmr teacher deployed item will reflect on teachers report

+++ app/Http/Livewire/WasteItem.php
@@ -55,13 +55,19 @@ class WasteItem extends Component
             ->where('item_type', '!=', 'consumable')
             ->where('quantity', '>', 0)
             ->get();
-        if (count($this->deployed_data) == 0){
-            Receiver::find($this->receiver_id)->delete();
-        }
+
     }
 
     public function displayMoved(){
         $this->movedData = MovedItem::where('receiver', $this->receiver_name)->get();
+        if (count($this->deployed_data) == 0 and count($this->movedData) == 0){
+            Receiver::find($this->receiver_id)->delete();
+            $fa = BackupPrepare::where('receiver','=',$this->receiver_name)->get();
+            foreach ($fa as $fas){
+                BackupPrepare::find($fas->id)->delete();
+            }
+        }
+
     }
     public function clickMove(){
         $data = BackupPrepare::find($this->waste_id);
